refactor: move login src files to new module
Move the views/viewmodels to the login module, still missing some of the navigation logic
