+++ app-login/src/main/java/com/chesire/nekome/app/login/details/LoginStatus.kt
@@ -1,4 +1,4 @@
-package com.chesire.nekome.flow.login.details
+package com.chesire.nekome.app.login.details
 
 /**
  * Provides the status of a login request.
