+++ app-login/src/main/res/navigation/login_nav_graph.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/login_nav_graph"
+    tools:ignore="UnusedNavigation">
+    <!-- UnusedNavigation is used because this is reached as a nested graph -->
+
+</navigation>
