+++ app/src/main/java/com/chesire/nekome/injection/androidmodules/FragmentModule.kt
@@ -9,8 +9,8 @@ import com.chesire.nekome.app.series.list.anime.AnimeFragment
 import com.chesire.nekome.app.series.list.manga.MangaFragment
 import com.chesire.nekome.app.settings.oss.OssFragment
 import com.chesire.nekome.app.timeline.TimelineFragment
-import com.chesire.nekome.flow.login.details.DetailsFragment
-import com.chesire.nekome.flow.login.syncing.SyncingFragment
+import com.chesire.nekome.app.login.details.DetailsFragment
+import com.chesire.nekome.app.login.syncing.SyncingFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
