+++ app/src/main/java/com/chesire/nekome/injection/androidmodules/ViewModelModule.kt
@@ -10,8 +10,8 @@ import com.chesire.nekome.app.series.detail.SeriesDetailViewModel
 import com.chesire.nekome.app.series.list.SeriesListViewModel
 import com.chesire.nekome.core.viewmodel.ViewModelFactory
 import com.chesire.nekome.flow.ActivityViewModel
-import com.chesire.nekome.flow.login.details.DetailsViewModel
-import com.chesire.nekome.flow.login.syncing.SyncingViewModel
+import com.chesire.nekome.app.login.details.DetailsViewModel
+import com.chesire.nekome.app.login.syncing.SyncingViewModel
 import dagger.Binds
 import dagger.MapKey
 import dagger.Module
