+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,13 +7,14 @@
 
     <!-- Nested -->
     <include app:graph="@navigation/discover_nav_graph" />
+    <include app:graph="@navigation/login_nav_graph" />
     <include app:graph="@navigation/search_nav_graph" />
     <include app:graph="@navigation/settings_nav_graph" />
 
     <!-- Login -->
     <fragment
         android:id="@+id/detailsFragment"
-        android:name="com.chesire.nekome.flow.login.details.DetailsFragment"
+        android:name="com.chesire.nekome.app.login.details.DetailsFragment"
         android:label="@string/nav_details"
         tools:layout="@layout/fragment_details">
         <action
@@ -22,7 +23,7 @@
     </fragment>
     <fragment
         android:id="@+id/syncingFragment"
-        android:name="com.chesire.nekome.flow.login.syncing.SyncingFragment"
+        android:name="com.chesire.nekome.app.login.syncing.SyncingFragment"
         android:label="@string/nav_syncing"
         tools:layout="@layout/fragment_syncing">
         <action
